Backend/backend/requirements.txt
django==5.2.8
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.9.0
psycopg2-binary==2.9.11
python-dotenv==1.2.1
Pillow==12.0.0
whitenoise==6.11.0
requests==2.32.5
gunicorn==21.2.0
firebase-admin==6.5.0
google-genai==0.3.0
python-decouple==3.8
dj-database-url==3.1.0
supabase==2.27.1